Move matrix free classes from CUDAWrappers namespace to Portable namespace
