+++ tests/matrix_free_kokkos/coefficient_eval_device.cc
@@ -24,7 +24,7 @@
 #include <deal.II/lac/cuda_vector.h>
 #include <deal.II/lac/la_parallel_vector.h>
 
-#include <deal.II/matrix_free/cuda_matrix_free.templates.h>
+#include <deal.II/matrix_free/portable_matrix_free.templates.h>
 
 #include "../tests.h"
 
@@ -38,12 +38,11 @@ public:
   DummyOperator() = default;
 
   DEAL_II_HOST_DEVICE void
-  operator()(
-    const unsigned int                                          cell,
-    const typename CUDAWrappers::MatrixFree<dim, double>::Data *gpu_data,
-    CUDAWrappers::SharedData<dim, double>                      *shared_data,
-    const double                                               *src,
-    double                                                     *dst) const;
+  operator()(const unsigned int                                      cell,
+             const typename Portable::MatrixFree<dim, double>::Data *gpu_data,
+             Portable::SharedData<dim, double> *shared_data,
+             const double                      *src,
+             double                            *dst) const;
 
   static const unsigned int n_dofs_1d = fe_degree + 1;
   static const unsigned int n_local_dofs =
@@ -57,9 +56,9 @@ public:
 template <int dim, int fe_degree>
 DEAL_II_HOST_DEVICE void
 DummyOperator<dim, fe_degree>::operator()(
-  const unsigned int                                          cell,
-  const typename CUDAWrappers::MatrixFree<dim, double>::Data *gpu_data,
-  CUDAWrappers::SharedData<dim, double>                      *shared_data,
+  const unsigned int                                      cell,
+  const typename Portable::MatrixFree<dim, double>::Data *gpu_data,
+  Portable::SharedData<dim, double>                      *shared_data,
   const double *,
   double *dst) const
 {
@@ -81,20 +80,20 @@ template <int dim, int fe_degree>
 class DummyMatrixFree : public Subscriptor
 {
 public:
-  DummyMatrixFree(const CUDAWrappers::MatrixFree<dim, double> &data_in,
-                  const unsigned int                           size);
+  DummyMatrixFree(const Portable::MatrixFree<dim, double> &data_in,
+                  const unsigned int                       size);
   void
   eval(LinearAlgebra::distributed::Vector<double, MemorySpace::Default> &dst)
     const;
 
 private:
-  const CUDAWrappers::MatrixFree<dim, double> &data;
+  const Portable::MatrixFree<dim, double> &data;
 };
 
 template <int dim, int fe_degree>
 DummyMatrixFree<dim, fe_degree>::DummyMatrixFree(
-  const CUDAWrappers::MatrixFree<dim, double> &data_in,
-  const unsigned int                           size)
+  const Portable::MatrixFree<dim, double> &data_in,
+  const unsigned int                       size)
   : data(data_in)
 {}
 
@@ -124,10 +123,9 @@ test()
   constraints.close();
 
   // Computation on the device
-  MappingQ<dim>                         mapping(fe_degree);
-  CUDAWrappers::MatrixFree<dim, double> mf_data;
-  typename CUDAWrappers::MatrixFree<dim, double>::AdditionalData
-    additional_data;
+  MappingQ<dim>                                              mapping(fe_degree);
+  Portable::MatrixFree<dim, double>                          mf_data;
+  typename Portable::MatrixFree<dim, double>::AdditionalData additional_data;
   additional_data.mapping_update_flags = update_values | update_gradients |
                                          update_JxW_values |
                                          update_quadrature_points;
@@ -154,10 +152,10 @@ test()
   const unsigned int n_colors = graph.size();
   for (unsigned int color = 0; color < n_colors; ++color)
     {
-      typename CUDAWrappers::MatrixFree<dim, double>::Data gpu_data =
+      typename Portable::MatrixFree<dim, double>::Data gpu_data =
         mf_data.get_data(color);
       const unsigned int n_cells = gpu_data.n_cells;
-      auto gpu_data_host = CUDAWrappers::copy_mf_data_to_host<dim, double>(
+      auto gpu_data_host         = Portable::copy_mf_data_to_host<dim, double>(
         gpu_data, additional_data.mapping_update_flags);
       for (unsigned int cell_id = 0; cell_id < n_cells; ++cell_id)
         {
