+++ tests/matrix_free_kokkos/matrix_free_device_initialize_vector.cc
@@ -15,7 +15,7 @@
 
 
 
-// Test CUDAWrappers::MatrixFree::initialize_dof_vector.
+// Test Portable::MatrixFree::initialize_dof_vector.
 
 #include <deal.II/distributed/tria.h>
 
@@ -29,7 +29,7 @@
 #include <deal.II/lac/affine_constraints.h>
 #include <deal.II/lac/cuda_vector.h>
 
-#include <deal.II/matrix_free/cuda_matrix_free.h>
+#include <deal.II/matrix_free/portable_matrix_free.h>
 
 #include <deal.II/numerics/vector_tools.h>
 
@@ -86,11 +86,10 @@ test()
   AffineConstraints<double> constraints(relevant_set);
   constraints.close();
 
-  MappingQ<dim>                         mapping(fe_degree);
-  CUDAWrappers::MatrixFree<dim, Number> mf_data;
-  const QGauss<1>                       quad(fe_degree + 1);
-  typename CUDAWrappers::MatrixFree<dim, Number>::AdditionalData
-    additional_data;
+  MappingQ<dim>                     mapping(fe_degree);
+  Portable::MatrixFree<dim, Number> mf_data;
+  const QGauss<1>                   quad(fe_degree + 1);
+  typename Portable::MatrixFree<dim, Number>::AdditionalData additional_data;
   mf_data.reinit(mapping, dof, constraints, quad, additional_data);
 
   VectorType vector;
