+++ tests/matrix_free_kokkos/matrix_vector_device_common.h
@@ -85,10 +85,9 @@ do_test(const DoFHandler<dim>           &dof,
 {
   deallog << "Testing " << dof.get_fe().get_name() << std::endl;
 
-  MappingQ<dim>                         mapping(fe_degree);
-  CUDAWrappers::MatrixFree<dim, Number> mf_data;
-  typename CUDAWrappers::MatrixFree<dim, Number>::AdditionalData
-    additional_data;
+  MappingQ<dim>                                              mapping(fe_degree);
+  Portable::MatrixFree<dim, Number>                          mf_data;
+  typename Portable::MatrixFree<dim, Number>::AdditionalData additional_data;
   additional_data.mapping_update_flags = update_values | update_gradients |
                                          update_JxW_values |
                                          update_quadrature_points;
